Facebook Poster/requirements.txt
requests
requests-oauthlib
huggingface_hub
schedule
python-dotenv
flask
